Bug fixed changed `change` eventListener to `input` on no_of_authors.
Previously the input text field with id `no_of_authors`, which was dynamically
added, had `change` eventlistener on it which forced the user to unfocus or
click outside the textfield to added more dynamic content. Some user without
knowing submitted the form as there was no content to be field which resulted
in error as the content dynamically added as per the input of `no_of_authors`
is compulsory to fill, So for better UX the eventListener on `no_of_authors`
was changed to `input` which added content as soon as the user types.